block_synchronizer: add new terminal state
`HaveStrictFinality` was considered the terminal state when
synchronizing a block even though this was not entirely true. Because on
the forward sync flow the synchronizer needed to wait until the block
was executed to establish that the sync was finished, there was some
unnecessary complexity added to keep `HaveStrictFinality` as a terminal
state which involved more complex checks in other intermediary states
like `HaveBlock`, `HaveAllDeploys` etc. The same can be said about the
historical sync flow where the synchronizer needed to wait for the block
to be marked complete.
This commit introduces a new terminal block acquisition state called
`Complete` which fully denotes that the block synchronizer is finished
with all tasks involving the block, on both forward and historical sync.
This simplifies some state transitions since enqueuing the block for
execution or requesting the block to be marked complete is only done
from the `HaveStrictFinality` state. For the forward sync case, there is
another intermediary state called `HaveFinalizedBlock`; the block
builder transitions to this state after the block has been successfully
converted into a block that is ready for execution.

Signed-off-by: Alexandru Sardan <alexandru@casperlabs.io>